Use getChildren() for the top-level element (Simulation) instead of invoking SimulationWrapper.getChildren() directly.  Return true for hasChildren based on what getChildren() returns, rather than if the element has property descriptors.
Store PropertySourceWrappers in an object-keyed hash so that the same PropertySourceWrapper gets used for each object in the view.  In addition to the PropertySourceWrapper, a list of its parentWrappers are also stored, allow for recursion to be detected.  If element A is a child of element B (anywhere), then B will not be shown as a child of B (anywhere).  This will allow for something to walk the tree arbitrarily deeply without fear of never ending (useful for bug 72)